fix: move terminal dialog to parent to survive context menu unmount
Clicking 'Open Terminal' in the context menu called onClose() which
set contextMenu to null, unmounting the HostNodeContextMenu component
before setUsernameDialogOpen(true) could take effect.

Moved the username dialog into a new HostNodeTerminalDialog component
rendered at the parent graph level, independent of the context menu
lifecycle. The context menu now emits onOpenTerminal(data) instead of
managing the dialog itself.
